chat_backend/requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.0.0
constantly==23.10.4
cryptography==44.0.1
daphne==4.0.0
Django==5.1.6
django-cors-headers==4.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
service-identity==24.2.0
setuptools==75.8.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
zope.interface==7.2
